Convert from ulong/ushort to stdint types in shared/.

+++ shared/mystrncpy.c
@@ -1,8 +1,10 @@
-#include "shared/types.h"
 #include <string.h>
 #include <stdlib.h>
+#include <stdint.h>
 
-void mystrncpy(uchar *dest,uchar *src,ulong len)
+#include "shared/types.h"
+
+void mystrncpy(uchar *dest,uchar *src,uint32_t len)
 {
    strncpy(dest,src,(size_t)len-1);
    dest[len-1]=0;
